Refactor shared test functionality and repackage to more clearly show that tests relate to akka streams
