+++ money-akka/src/test/scala/com/comcast/money/akka/Blocking.scala
@@ -1,0 +1,12 @@
+package com.comcast.money.akka
+
+import scala.concurrent.duration.DurationLong
+import scala.concurrent.{Await, Future}
+
+object Blocking {
+
+  implicit class RichFuture[T](future: Future[T]) {
+    def get: T = Await.result(future, 5 seconds)
+  }
+
+}
