+++ money-akka/src/test/scala/com/comcast/money/akka/CollectingSpanHandler.scala
@@ -1,0 +1,15 @@
+package com.comcast.money.akka
+
+import com.comcast.money.api.{SpanHandler, SpanInfo}
+
+class CollectingSpanHandler() extends SpanHandler {
+  var spanInfoStack = Seq.empty[SpanInfo]
+
+  def push(span: SpanInfo): Unit = spanInfoStack = span +: spanInfoStack
+
+  def pop: Option[SpanInfo] = spanInfoStack.headOption
+
+  def clear(): Unit = spanInfoStack = Seq.empty[SpanInfo]
+
+  override def handle(span: SpanInfo): Unit = push(span)
+}
